refactor: Refactor `ddev get` into sub-commands, fixes #6146
