+++ cmd/ddev/cmd/testdata/TestCmdAddon/ddev-memcached.tar.gz
