+++ cmd/ddev/cmd/testdata/TestCmdAddon/example-repo/docker-compose.example.yaml
