+++ cmd/ddev/cmd/testdata/TestCmdAddon/example-repo/file-with-no-ddev-generated.txt
