+++ cmd/ddev/cmd/testdata/TestCmdAddon/example-repo/globalextras/nosig.txt
