+++ cmd/ddev/cmd/testdata/TestCmdAddon/example-repo/globalextras/okfile.txt
