+++ cmd/ddev/cmd/testdata/TestCmdAddon/example-repo/install.yaml
