+++ cmd/ddev/cmd/testdata/TestCmdAddon/example-repo/projectextras/somefiles/nosig.txt
