+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/project/.platform.app.yaml
