+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/project/.platform/local/.gitignore
