+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/project/.platform/local/README.txt
