+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/project/.platform/routes.yaml
