+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/project/.platform/services.yaml
