+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/recipe/extra/has-ddev-generated.txt
