+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/recipe/extra/no-ddev-generated.txt
