+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/recipe/install.yaml
