+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/recipe/junk_darwin_amd64.txt
