+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/recipe/junk_darwin_arm64.txt
