+++ cmd/ddev/cmd/testdata/TestCmdAddonComplex/recipe/junk_windows_amd64.txt
