+++ cmd/ddev/cmd/testdata/TestCmdAddonDependencies/dependency_recipe/install.yaml
