+++ docs/content/users/usage/architecture.md
@@ -20,7 +20,7 @@ A project’s `.ddev` directory can be intimidating at first, so let’s take a 
     For example, if you see a `solr` directory, it probably pertains to a custom Solr [add-on service](../extend/additional-services.md).
 
 `addon-metadata` directory
-: Contains metadata about add-on services that have been added to the project. This allows commands like `ddev get --installed` and `ddev get --remove` to work, see [Managing Installed Add-Ons](../extend/additional-services.md#managing-installed-add-ons).
+: Contains metadata about add-on services that have been added to the project. This allows commands like `ddev add-on list --installed` and `ddev add-on remove` to work, see [Managing Installed Add-Ons](../extend/additional-services.md#managing-installed-add-ons).
 
 `apache` directory
 : Default Apache configuration when using `webserver_type: apache-fpm`, which [can be customized](../extend/customization-extendibility.md#custom-apache-configuration).
