+++ docs/content/users/usage/commands.md
@@ -614,66 +614,89 @@ ddev export-db > /tmp/db.sql.gz
 ddev export-db my-project --gzip=false --file=/tmp/my-project.sql
 ```
 
-## `get`
+## `add-on`
 
-Download an [add-on](../extend/additional-services.md) (service, provider, etc.).
+[Add-on](../extend/additional-services.md) commands.
+
+Environment variables:
+
+* `DDEV_GITHUB_TOKEN`: A [GitHub token](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token) may be used for `ddev add-on` requests (which result in GitHub API queries). It's unusual for casual users to need this, but if you're doing lots of `ddev add-on` requests you may run into rate limiting. The token you use requires no privileges at all. Example:
+
+```bash
+export DDEV_GITHUB_TOKEN=<your github token>
+ddev add-on list --all
+```
+
+### `add-on get`
+
+Download an add-on (service, provider, etc.).
+
+Flags:
+
+* `--version <version>`: Specify a version to download
+* `--verbose`, `-v`: Output verbose error information with Bash `set -x` (default `false`)
+
+Example:
+
+```shell
+# Download the official Redis add-on
+ddev add-on get ddev/ddev-redis
+
+# Get debug info about `ddev add-on get` failure
+ddev add-on get ddev/ddev-redis --verbose
+
+# Download the official Redis add-on, version v1.0.4
+ddev add-on get ddev/ddev-redis --version v1.0.4
+
+# Download the Drupal Solr add-on from its v1.2.3 release tarball
+ddev add-on get https://github.com/ddev/ddev-drupal-solr/archive/refs/tags/v1.2.3.tar.gz
+
+# Copy an add-on available in another directory
+ddev add-on get /path/to/package
+
+# Copy an add-on from a tarball in another directory
+ddev add-on get /path/to/tarball.tar.gz
+```
+
+In general, you can run `ddev add-on get` multiple times without doing any damage. Updating an add-on can be done by running `ddev add-on get <add-on-name>`. If you have changed an add-on file and removed the `#ddev-generated` marker in the file, that file will not be touched and DDEV will let you know about it.
+
+### `add-on remove`
+
+Remove an installed add-on. Accepts the full add-on name, the short name of the repository, or with owner/repository format.
+
+Flags:
+
+* `--verbose`, `-v`: Output verbose error information with Bash `set -x` (default `false`)
+
+Example:
+
+```shell
+ddev add-on remove redis
+ddev add-on remove ddev-redis
+ddev add-on remove ddev/ddev-redis
+```
+
+### `add-on list`
+
+Download an add-on (service, provider, etc.).
 
 Flags:
 
 * `--all`: List unofficial *and* official add-ons. (default `true`)
-* `--list`: List official add-ons. (default `true`)
 * `--installed`: List installed add-ons
-* `--remove <add-on>`: Remove an installed add-on
-* `--version <version>`: Specify a version to download
-* `--verbose`, `-v`: Output verbose error information with Bash `set -x` (default `false`)
-
-Environment variables:
-
-* `DDEV_GITHUB_TOKEN`: A [GitHub token](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token) may be used for `ddev get` requests (which result in GitHub API queries). It's unusual for users to need this, but if you're doing lots of `ddev get` requests you may run into rate limiting. The token you use requires no privileges at all. Example:
-
-```bash
-export DDEV_GITHUB_TOKEN=<your github token>
-ddev get --list --all
-```
 
 Example:
 
 ```shell
 # List official add-ons
-ddev get --list
+ddev add-on list
 
 # List official and third-party add-ons
-ddev get --list --all
+ddev add-on list --all
 
-# Download the official Redis add-on
-ddev get ddev/ddev-redis
-
-# Get debug info about `ddev get` failure
-ddev get ddev/ddev-redis --verbose
-
-# Download the official Redis add-on, version v1.0.4
-ddev get ddev/ddev-redis --version v1.0.4
-
-# Download the Drupal Solr add-on from its v1.2.3 release tarball
-ddev get https://github.com/ddev/ddev-drupal-solr/archive/refs/tags/v1.2.3.tar.gz
-
-# Copy an add-on available in another directory
-ddev get /path/to/package
-
-# Copy an add-on from a tarball in another directory
-ddev get /path/to/tarball.tar.gz
-
-# View installed add-ons
-ddev get --installed
-
-# Remove an add-on can be done with the full name, the short name of repo
-# or with owner/repo format
-ddev get --remove redis
-ddev get --remove ddev-redis
-ddev get --remove ddev/ddev-redis
+# List installed add-ons
+ddev add-on list --installed
 ```
-
-In general, you can run `ddev get` multiple times without doing any damage. Updating an add-on can be done by running `ddev get <add-on-name>`. If you have changed an add-on file and removed the `#ddev-generated` marker in the file, that file will not be touched and DDEV will let you know about it.
 
 ## `heidisql`
