+++ pkg/ddevapp/dotddev_assets/providers/platform.yaml
@@ -3,7 +3,7 @@
 # your own preferences. If you edit it, remove the `ddev-generated` line from the top so
 # that it won't be overwritten.
 
-# Consider using `ddev get ddev/ddev-platformsh` (https://github.com/ddev/ddev-platformsh) for more
+# Consider using `ddev add-on get ddev/ddev-platformsh` (https://github.com/ddev/ddev-platformsh) for more
 # complete platform integration.
 
 # To use this configuration,
