+++ pkg/ddevapp/global_dotddev_assets/commands/host/launch
@@ -42,7 +42,7 @@ replace_port_in_cloud() {
 while :; do
   case ${1:-} in
   -p | --phpmyadmin)
-    echo "phpMyAdmin is no longer built into DDEV, please 'ddev get ddev/ddev-phpmyadmin' and use 'ddev phpmyadmin' to launch phpMyAdmin" && exit 2
+    echo "phpMyAdmin is no longer built into DDEV, please 'ddev add-on get ddev/ddev-phpmyadmin' and use 'ddev phpmyadmin' to launch phpMyAdmin" && exit 2
     ;;
   -m | --mailpit | --mailhog)
     if [ "${IN_CLOUD}" != "" ]; then
